Updated Thing.go_to_state to be a private function. It should never be invoked directly. Move fixture classes to their own files so they can be shared by multiple test scripts.
